Add files via upload
The above Python file (made for use with Jupyter Books) is an easy method of installing datasets from Kaggle for use to be easily used for analytics or querying using Python's opendatasets, os and pandas packages. This file has instructions for use. Feel free to paste this in for easy dataset import for your personal projects!